+++ src/refactorings/extract/extract-generic-type/extract-generic-type.ts
@@ -83,6 +83,7 @@ function createVisitor(
       if (!t.isSelectablePath(path)) return;
 
       const interfaceDeclaration = findParentInterfaceDeclaration(path);
+      const functionDeclaration = findParentFunctionDeclaration(path);
 
       if (interfaceDeclaration) {
         if (!selection.isInsidePath(interfaceDeclaration)) return;
@@ -91,6 +92,13 @@ function createVisitor(
         if (!selection.isInsidePath(path)) return;
 
         onMatch(new SelectedInterfaceOccurrence(path, interfaceDeclaration));
+      } else if (functionDeclaration) {
+        if (!selection.isInsidePath(functionDeclaration)) return;
+
+        onVisit(new FunctionOccurrence(path, functionDeclaration));
+        if (!selection.isInsidePath(path)) return;
+
+        onMatch(new SelectedFunctionOccurrence(path, functionDeclaration));
       }
     }
   };
@@ -101,6 +109,14 @@ function findParentInterfaceDeclaration(
 ) {
   return path.findParent(t.isTSInterfaceDeclaration) as t.NodePath<
     t.TSInterfaceDeclaration
+  > | null;
+}
+
+function findParentFunctionDeclaration(
+  path: t.SelectablePath<t.TSTypeAnnotation>
+) {
+  return path.findParent(t.isFunctionDeclaration) as t.NodePath<
+    t.FunctionDeclaration
   > | null;
 }
 
@@ -196,6 +212,99 @@ class SelectedInterfaceOccurrence extends InterfaceOccurrence {
     );
 
     this.interfaceDeclaration.node.typeParameters = t.tsTypeParameterDeclaration(
+      [...this.existingTypeParameters, newTypeParameter]
+    );
+  }
+}
+
+class FunctionOccurrence implements Occurrence {
+  readonly symbolPosition?: Position;
+  protected readonly typeName: string;
+
+  constructor(
+    readonly path: t.SelectablePath<t.TSTypeAnnotation>,
+    protected functionDeclaration: t.NodePath<t.FunctionDeclaration>
+  ) {
+    this.symbolPosition = this.determineSymbolPosition();
+    this.typeName = this.computeValidTypeName();
+  }
+
+  get node(): t.Selectable<t.TSTypeAnnotation> {
+    return this.path.node;
+  }
+
+  transform() {
+    const typeAnnotation = t.tsTypeAnnotation(
+      t.tsTypeReference(t.identifier(this.typeName))
+    );
+    this.path.replaceWith(typeAnnotation);
+  }
+
+  protected get existingTypeParameters(): t.TSTypeParameter[] {
+    const NO_PARAMS: t.TSTypeParameter[] = [];
+    const { typeParameters } = this.functionDeclaration.node;
+    return (
+      (t.isTSTypeParameterDeclaration(typeParameters) &&
+        typeParameters.params) ||
+      NO_PARAMS
+    );
+  }
+
+  private determineSymbolPosition(): Position | undefined {
+    const lastTypeParameter = last(this.existingTypeParameters);
+    if (lastTypeParameter) {
+      if (!t.isSelectableNode(lastTypeParameter)) return;
+
+      /**
+       * function position<T = number>() {
+       *                            ^−−−− end of last type param
+       *
+       * function position<T = number, U = string>() {
+       *                              ^−−−− end of last type param + 2
+       */
+      return Position.fromAST(lastTypeParameter.loc.end).addCharacters(2);
+    } else {
+      const { id } = this.functionDeclaration.node;
+      if (!t.isSelectableNode(id)) return;
+
+      /**
+       * function position() {
+       *                 ^−−−− end of ID
+       *
+       * function position<T = number>() {
+       *                  ^−−−− end of ID + 1
+       */
+      return Position.fromAST(id.loc.end).addCharacters(1);
+    }
+  }
+
+  private computeValidTypeName(): string {
+    const DEFAULT_NAME = "T";
+    const VALID_NAMES = [DEFAULT_NAME, "U", "V", "W", "X", "Y", "Z"];
+
+    const existingNames = this.existingTypeParameters.map(({ name }) => name);
+    const availableNames = VALID_NAMES.filter(
+      name => !existingNames.includes(name)
+    );
+
+    return availableNames[0] || DEFAULT_NAME;
+  }
+}
+
+class SelectedFunctionOccurrence extends FunctionOccurrence {
+  transform() {
+    this.addGenericDeclaration();
+    super.transform();
+  }
+
+  private addGenericDeclaration() {
+    const newTypeParameter = t.tsTypeParameter(
+      undefined,
+      this.path.node.typeAnnotation,
+      this.typeName
+    );
+
+    this.functionDeclaration.node.typeParameters = t.tsTypeParameterDeclaration(
       [...this.existingTypeParameters, newTypeParameter]
     );
   }