+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- "Extract Generic Type" now works on function declarations.
+
 ## [4.2.0] - 2020-05-01 - Make it Generic 🏭
 
 ### Added